more debugging tools
