+++ software/code/src/build_test.sh
@@ -1,0 +1,1 @@
+g++ test.cpp Cartridge.cpp GPIO.cpp -lwiringPi -o test.out
