+++ tests/roms/.gitignore
@@ -1,0 +1,4 @@
+*.gb
+*.gba
+*.gbc
+*.out
