+++ tests/roms/build_4000steps.sh
@@ -1,0 +1,1 @@
+gcc 4000steps.c -o 4000steps.out
